Broke the program up into modules
Looks like a real C project now, if you didn't know better, it might
look like I might know what I'm doing.
