+++ board.h
@@ -1,0 +1,7 @@
+#ifndef BOARD_H
+#define BOARD_H
+
+extern void updatescore ();
+extern void updateboard ();
+
+#endif
