+++ boardedit.h
@@ -1,0 +1,7 @@
+#ifndef BOARDEDIT_H
+#define BOARDEDIT_H
+
+extern void markcol (unsigned char mx,unsigned char my);
+extern void markrow (unsigned char mx,unsigned char my);
+
+#endif
