+++ button.h
@@ -1,0 +1,8 @@
+#ifndef BUTTON_H
+#define BUTTON_H
+
+#include "file.h"
+
+extern void mbutton (unsigned char mx, unsigned char my);
+
+#endif
