+++ draw.h
@@ -1,0 +1,8 @@
+#ifndef DRAW_H
+#define DRAW_H
+
+extern void cboxclear (unsigned char tx,unsigned char ty,unsigned char bx,unsigned char by);
+extern void cbox (unsigned char tx,unsigned char ty,unsigned char bx,unsigned char by);
+extern void drawbutton (char bx, char by, char btext[]);
+
+#endif
