+++ gridgame.c
@@ -25,41 +25,15 @@
 #include <string.h>
 #include <stdbool.h>
 
-#define LX  12	// low 'x' val (left side of playfield)
-#define HX  27	// high 'x' val (right side of playfield)
-#define LY  1	// low 'y' val (top of playfield)
-#define HY  16	// high 'y' val (bottom of playfield)
+#include "global.h"
+#include "button.h"
+#include "draw.h"
+#include "board.h"
+#include "boardedit.h"
 
 // Some colors we use in here
 #define WHITE 97
 #define RED   98
-
-unsigned char filename[16];
-unsigned char tmpname[20];
-static char fbuf[256];
-static char board[16][16];
-unsigned int score;
-unsigned int hscore;
-bool overwrite;
-bool nhs;
-bool editmode;
-bool sb;
-
-bool redrawboard = 1;
-
-unsigned char grid[4][8] = {
-	{ 117,96,105,160,160,160,160,98 },
-	{ 98,160,160,160,160,113,160,98 },
-	{ 98,96,105,117,96,98,117,179 },
-	{ 106,96,107,98,160,98,106,107 }
-};
-
-unsigned char game[4][9] = {
-	{ 117,105,117,105,160,160,160,117,105 },
-	{ 106,179,117,179,117,178,105,171,107 },
-	{ 160,98,106,107,98,98,98,106,107 },
-	{ 106,107,160,160,160,160,160,160,160 }
-};
 
 unsigned char dirtable[4][2] = {
 	{ 117 , 106 },
@@ -114,6 +88,7 @@ void Enqueue (Queue *Q,unsigned char ex, unsigned char ey) {
 		++Q->size;
 		++Q->rear;
 		// Should probably just make the Q bigger, this makes the Q circular
+		// Pretty sure this is one of the places things go goofy on the C-64 version
 		if (Q->rear == Q->capacity) {
 			Q->rear = 0;
 		}
@@ -124,329 +99,10 @@ void Enqueue (Queue *Q,unsigned char ex, unsigned char ey) {
 	return;
 }
 
-void cboxclear (unsigned char tx,unsigned char ty,unsigned char bx,unsigned char by) {
-	unsigned char bw,bh;
-	unsigned char y;
-
-	bw = (bx - tx) + 1;
-	bh = (by - ty) + 1;
-	
-	for (y = ty; y <= by; ++y) {
-		cclearxy (tx,y,bw);
-	}
-}
-
-void cbox (unsigned char tx,unsigned char ty,unsigned char bx,unsigned char by) {
-	unsigned char bw,bh;
-	unsigned char hlc,vlc; // horizontal/vertical line drawing char
-	unsigned char x;
-	
-	bw = (bx - tx) - 1;
-	bh = (by - ty) - 1;
-	
-	if (editmode) {
-		hlc = 42;  // '*'
-		vlc = hlc;
-	} else {
-		hlc = 96;  // shifted 'C'
-		vlc = 98;  // shifted 'B'
-	}
-	
-	cputcxy (tx,ty,117);   // TL corner
-	cputcxy (bx,ty,105);   // TR corner
-	cputcxy (bx,by,107);   // BR corner
-	cputcxy (tx,by,106);   // BL corner
-	
-	for (x = 1; x <= bw; ++x) {
-		cputcxy (tx+x,ty,hlc);
-		cputcxy (tx+x,by,hlc);
-	}
-	
-	for (x = 1; x <= bh; ++x) {
-		cputcxy (tx,ty+x,vlc);
-		cputcxy (bx,ty+x,vlc);
-	}
-}
-
-void drawbutton (char bx, char by, char btext[]) {
-	unsigned char bl = strlen (btext);
-	bool em = editmode; // save the editmode
-
-	editmode = 0; // we want real line draw chars, not '*'
-	
-	cbox (bx,by,bx+bl+1,by+2);
-	cputsxy (bx+1,by+1,btext);
-	
-	editmode = em; // back to whatever it was before
-}
-
-void printboard () {
-	unsigned char x,y;
-
-	clrscr ();
-	gotoxy (0,0);
-	
-	for (y = 0; y <= 15; ++y) {
-		for (x = 0; x <= 15; ++x) {
-			cprintf ("%c",board[x][y]);
-		}
-		cprintf ("\r\n");
-	}
-	cgetc ();
-}
-
-// TODO: handle non-existing file for board loading
-void loadfile (unsigned char filename[], unsigned char fsize) {
-	unsigned char y;
-	
-	if (overwrite == 1) {
-		strcpy (tmpname,filename);
-		strcat (tmpname,",s,r");
-
-		cbm_open (2,8,2,tmpname);
-		
-		if (sb) {
-			for (y = 0; y <= 15; ++y) {
-				fbuf[0] = '\0';
-				gotoxy (LX,LY + y);
-				cbm_read (2,fbuf,16); // only read 16 chars
-				cprintf ("%s",fbuf);
-			}
-			sb = 0;
-		} else { 
-			cbm_read (2,fbuf,fsize);
-			cbm_close (2);
-		}
-	}
-}
-
-void append (char* s,char c) {
-	unsigned char len;
-
-	len = strlen(s);
-	
-	s[len] = c;
-	s[len + 1] = '\0';
-}
-
-// TODO: handle existing file for board saving
-void savefile (unsigned char filename[], unsigned int fsize) {
-	unsigned char x,y;
-	
-	if (overwrite == 1) {
-		strcpy (tmpname,"s0:");
-		strcat (tmpname,filename);
-		strcat (tmpname,",s");
-		
-		cbm_open (2,8,15,tmpname);
-		cbm_close (2);
-	}
-
-	strcpy (tmpname,filename);
-	strcat (tmpname,",s,w");
-	cbm_open (2,8,2,tmpname);
-	
-	if (sb) {
-		for (y = 0; y <= 15; ++y) {
-			fbuf[0] = '\0';
-			for (x = 0; x <= 15; ++x) {
-				append (fbuf,board[x][y]);
-			}
-			// cprintf ("%s\r\n",fbuf);
-			cbm_write (2,fbuf,16); // only write 16 chars
-		}
-		sb = 0;
-	} else {
-		cbm_write (2,fbuf,fsize);
-	}
-
-	cbm_close (2);
-}
-
-void updatescore () {
-	if (score >= hscore) {
-		nhs = true;
-		hscore = score;
-	}
-	if (hscore || nhs == true) {
-		gotoxy (2,10);
-		cprintf ("%6d",hscore);
-	}
-	
-	gotoxy (2,6);
-	cprintf ("%6d",score);
-}
-
-void loadhs () {
-	overwrite = 1;
-	loadfile ("highscore",6);  // highscores are max 6 chars long.
-	hscore = atoi (fbuf);
-}
-
-void savehs () {
-	if (editmode) {
-		nhs = true;
-		hscore = 0;
-		updatescore ();
-	}
-	
-	if (nhs == true) {
-		overwrite = 1;
-		gotoxy (0,24);
-		itoa (hscore,fbuf,10);
-		gotoxy (0,24);
-		savefile ("highscore",strlen (fbuf));
-		nhs = false;
-	}
-}
-
-void updateboard () {
-	unsigned char row,rows,cols,x;
-	unsigned char cpart;
-
-	// print the "Grid" text
-	row = 0;
-	gotoxy (2,row);
-	for (rows = 0; rows < 4; ++rows) {
-		for (cols = 0; cols < 8; ++cols) {
-			cprintf ("%c",grid[rows][cols]);
-		}
-		row++;
-		gotoxy (2,row);
-	}
-
-	// print the "game" text
-	row = 0;
-	gotoxy (29,row);
-	for (rows = 0; rows < 4; ++rows) {
-		for (cols = 0; cols < 9; ++cols) {
-			cprintf ("%c",game[rows][cols]);
-		}
-		row++;
-		gotoxy (29,row);
-	}
-
-	cbox (LX-1,LY-1,HX+1,HY+1);
-
-	if (redrawboard !=0) {
-		for (rows = LY; rows <= HY ; ++rows) {
-			for (cols = LX; cols <= HX ; ++cols) {
-				x = (rand () % 4);
-				switch (x) {
-					case 0:
-						cpart = 105;
-						break;
-					case 1:
-						cpart = 106;
-						break;
-					case 2:
-						cpart = 107;
-						break;
-					case 3:
-						cpart = 117;
-						break;
-				}
-				cputcxy (cols,rows,cpart);
-			}
-		}
-	}
-	
-	cputsxy (2,5,"score:");
-	cputsxy (2,8,"high");
-	cputsxy (2,9,"score:");
-	
-	updatescore ();
-	
-	if (editmode) {
-		cboxclear (31,5,37,7);      // erase 'reset' button
-		cboxclear (0,18,39,24);     // erase the bottom of screen
-		drawbutton (31,8," done");  // want 'reset' and 'done' buttons to be same size
-		drawbutton (31,11," load"); // load new board
-		drawbutton (31,14," save"); // save board to disk
-		drawbutton (1,11,"reset");  // high score reset button
-		gotoxy(0,19);
-		cprintf("click on a piece to rotate it\r\n");
-		cprintf("clicking on an '*' on the border will\r\n");
-		cprintf("change the whole row or column to match\r\n");
-		cprintf("the adjacent piece.");
-	} else {
-		drawbutton (31,5,"reset");  // generate new random board
-		drawbutton (31,8," edit");  // edit board
-		cboxclear (31,11,37,16);    // remove the load/save buttons above 
-		cboxclear (1,11,7,13);      // remove the 'reset' button under the high score
-
-		cboxclear (0,18,39,24);     // erase the bottom of screen
-		gotoxy (0,19);
-		cprintf("the objective of the game is to get a\r\n");
-		cprintf("chain reaction of tiles as long as\r\n"); 
-		cprintf("possible.tiles will start each other if\r\n");
-		cprintf("their lines are connected.\r\n");
-		cprintf("click on a tile to start.");
-	}
-}
-
-void loadboard () {
-	unsigned char c;
-	
-	cursor (1);
-	cboxclear (0,18,39,24); // clear the bottom of the screen}
-	cputsxy (0,18,"filename? ");
-	scanf ("%s",&filename);
-	
-	gotoxy (0,19);
-	cprintf ("loading board: %s",filename);
-	overwrite = 1;          // force attempting to load (TODO: why?)
-	sb = 1;
-	loadfile (filename,1);   // board loading does it's own file sizing
-	cclearxy (0,21,40);      // clear 'load...' text.
-
-	gotoxy (0,21);
-	cprintf ("board %s loaded\r\n",filename);
-	// wait for 2 seconds
-	for (c = 0; c <= 120 ; ++c) { 
-		waitvsync ();
-	}
-	cboxclear (0,18,39,24); // clear the bottom of the screen}
-	cursor (0);
-	updateboard ();
-}
-
-void saveboard () {
-	unsigned char x,y;
-	unsigned char c;
-
-	cursor (1);
-	cputsxy (0,18,"filename? ");
-	scanf ("%s",&filename);
-	
-	for (y = 0; y <= 15; ++y) {
-		for (x = 0; x <= 15; ++x) {
-			gotoxy (x + LX,y + LY);
-			board[x][y] = cpeekc ();
-		}
-	}
-	
-	gotoxy (0,19);
-	cprintf ("saving board: %s",filename);
-	overwrite = 1; // force attempting to save
-	sb = 1; // so savefile () knows to do the board handling
-	savefile (filename,1);
-	cclearxy (0,21,40); // clear 'saving...' text.
-	gotoxy (0,21);
-	cprintf ("board %s saved\r\n",filename);
-	// wait for 2 seconds
-	for (c = 0; c <= 120 ; ++c) { 
-		waitvsync ();
-	}
-	cboxclear (0,18,39,24); // clear the bottom of the screen
-	cursor (0);
-	updateboard ();
-}
-
 void check (Queue *Q, unsigned char cx, unsigned char cy, unsigned char dir) {
 	unsigned char c,cc;
 
-	if (editmode) {
+	if (global_editmode) {
 		return;
 	}
 	
@@ -501,7 +157,7 @@ void check (Queue *Q, unsigned char cx, unsigned char cy, unsigned char dir) {
 		textcolor (WHITE);
 		
 		Enqueue (Q,cx,cy);
-		++score;
+		++global_score;
 		updatescore ();
 	}
 }
@@ -553,87 +209,6 @@ void processQ(Queue *Q) {
 	cputsxy (2,16,"ready");
 }
 
-void markcol (unsigned char mx,unsigned char my) {
-	unsigned char c,y;
-	
-	// clicked on top column marker
-	if (my == 0) {
-		gotoxy (mx,1);
-		c = cpeekc ();
-		
-	}
-	// clicked on bottom column marker
-	else if (my == 17) {
-		gotoxy (mx,16);
-		c = cpeekc ();
-	}
-
-	for (y = 1; y <= 16; ++y) {
-		cputcxy (mx,y,c);
-	}
-}
-
-void markrow (unsigned char mx,unsigned char my) {
-	unsigned char c,x;
-	
-	// clicked on top column marker
-	if (mx == 11) {
-		gotoxy (12,my);
-		c = cpeekc ();
-		
-	}
-	// clicked on bottom column marker
-	else if (mx == 28) {
-		gotoxy (27,my);
-		c = cpeekc ();
-	}
-
-	for (x = 12; x <= 27; ++x) {
-		cputcxy (x,my,c);
-	}
-}
-
-void mbutton (unsigned char mx, unsigned char my) {
-	// clicked on game 'reset' button
-	if ((mx >= 31 && mx <= 37) && (my >= 5 && my <= 7)) {
-		redrawboard = 1;
-		clrscr ();
-		updateboard ();
-	} 
-	// clicked on 'done/edit' button
-	else if ((mx >= 31 && mx <= 37) && (my >= 8 && my <= 10)) {
-		if (editmode) {
-			editmode = 0;
-			redrawboard = 0;
-			score = 0;
-			updateboard ();
-		} else {
-			editmode = 1;
-			redrawboard = 0;
-			updateboard ();
-		}
-	}
-	if (editmode) {
-		if ((mx >= 31 && mx <= 37) && (my >= 11 && my <= 13)) {
-			loadboard ();
-		}
-		if ((mx >= 31 && mx <= 37) && (my >= 14 && my <= 17)) {
-			saveboard ();
-		}
-		if ((mx >= 1 && mx <= 7) && (my >=11 && my <= 13)) {
-			savehs ();
-		}
-		if ((mx >= 12 && mx <= 27) && (my == 0 || my == 17)) {
-			// mark column same as top/bottom cell
-			markcol (mx,my);
-		}
-		if ((mx == 11 || mx == 28) && (my >= 1 && my <= 16)) {
-			// mark row same as left/right
-			markrow (mx,my);
-		}
-	}
-}
-
 int main () {
 	struct mouse_info info;
 	unsigned char c,mbl,mx,my;
@@ -658,7 +233,7 @@ int main () {
 	mouse_show ();
 
 	loadhs ();
-	
+	global_redrawboard = true;
 	updateboard ();
 	
 	while (1) {
@@ -675,7 +250,7 @@ int main () {
 			// check for clicks on gameboard
 			if ((mx >= LX && mx <= HX) && (my >= LY && my <= HY)) {
 				cclearxy (2,16,5);  // clear the 'ready' text 
-				score = 0;
+				global_score = 0;
 				updatescore ();
 				gotoxy (mx,my);
 				c = cpeekc ();
@@ -687,7 +262,7 @@ int main () {
 			} else {
 				mbutton (mx,my);
 			}
-			if (editmode == 0) {
+			if (global_editmode == 0) {
 				savehs ();
 			}
 		}
